Make packages, add documentation, minor refactoring
