+++ src/main/java/Accounts/Account.java
@@ -1,4 +1,7 @@
-package FlashcardProgram;
+package Accounts;
+
+import Decks.Deck;
+import Sessions.StudySession;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -9,7 +12,7 @@ public class Account {
     private String username;
     private final String password;
     private final List<Deck> decks;
-    private final Map<Deck, List<StudySession>> decksToSessions;
+    private final Map<Deck, List<StudySession>> decksToSessions; // maps decks to StudySessions that use them
 
     public Account(String username, String password, List<Deck> decks){
         this.username = username;
@@ -64,46 +67,5 @@ public class Account {
 
     public void deleteSession(Deck deck, StudySession session) {
         this.decksToSessions.get(deck).remove(session);
-    }
-
-    public void updateSessionsOfDeck(Deck deck) {
-
-        List<Flashcard> flashcardList = deck.getFlashcards();
-
-        List<StudySession> listOfSessions = this.decksToSessions.get(deck);
-
-        for (StudySession session : listOfSessions) {
-            // First, check for updates needed from adding a card:
-            // Loop through flashcardList. If it is in flashcardData, move on.
-            // If it is not in flashcardData, add it to flashcardData.
-            // Second, check for updates needed from deleting a card:
-            // Loop through flashcardData. If it is in flashcardList, move on.
-            // If it is not in flashcardList, then delete it from flashcardData
-
-            Map<Flashcard, FlashcardData> flashcardToFlashcardData = session.getFlashcardToData();
-
-            for (Flashcard flashcard : flashcardList) {
-                if (!session.getFlashcardToData().containsKey(flashcard)) {
-                    session.getFlashcardToData().put(flashcard, new FlashcardData(0));
-                }
-            }
-
-            for (Flashcard flashcard : flashcardToFlashcardData.keySet()) {
-                if (!flashcardList.contains(flashcard)) {
-                    flashcardToFlashcardData.remove(flashcard);
-                }
-            }
-
-            session.cardShuffler.updateDeckContext();
-
-        }
-
-
-
-
-
-
-
-
     }
 }
