+++ src/main/java/Main/Main.java
@@ -1,4 +1,4 @@
-package FlashcardProgram;
+package Main;
 
 public class Main {
     public static void main(String[] args) {
