+++ src/main/java/Sessions/PracticeSession.java
@@ -1,4 +1,8 @@
-package FlashcardProgram;
+package Sessions;
+
+import Decks.Deck;
+import Sessions.BasicShuffle;
+import Sessions.StudySession;
 
 public class PracticeSession extends StudySession {
     /**
